+++ src/app/(website)/register/page.tsx
@@ -1,14 +1,11 @@
 import { auth } from "@/auth";
-import { db } from "@/db/main";
-import { CategoryTable } from "@/db/schema";
+import CategoryOptionsCard from "@/components/forms/categories/category-options-card";
+import CategoryRegistrationCard from "@/components/forms/categories/category-registration-card";
+import ProductRegistrationCard from "@/components/forms/products/product-registration-card";
 import { redirect } from "next/navigation";
-import CategoryOptionsForm from "../products/(components)/category-options-form";
-import CategoryRegistrationForm from "../products/(components)/category-registration-form";
-import ProductRegistrationForm from "../products/(components)/product-registration-form";
 
 const Page = async () => {
   const session = await auth();
-  const categories = await db.select().from(CategoryTable).all();
 
   if (!session) redirect("/");
 
@@ -18,15 +15,20 @@ const Page = async () => {
         Crea tus productos
       </h1>
 
-      <section className="my-5 grid w-full max-w-3xl grid-cols-1 gap-3">
-        <CategoryRegistrationForm session={session} />
-        <CategoryOptionsForm categories={categories} />
+      <div className="my-5 grid grid-cols-1 gap-3 sm:grid-cols-2">
+        <section className="grid w-full grid-cols-1 gap-3">
+          <CategoryRegistrationCard />
 
-        <div className="flex w-full flex-col gap-3">
-          <ProductRegistrationForm session={session} />
-          {/* <ProductCategoryForm /> */}
-        </div>
-      </section>
+          <div className="flex w-full flex-col gap-3">
+            <ProductRegistrationCard />
+            {/* <ProductCategoryForm /> */}
+          </div>
+        </section>
+
+        <section>
+          <CategoryOptionsCard />
+        </section>
+      </div>
     </main>
   );
 };
