+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 import Footer from "@/components/footer/footer";
-import Header from "@/components/header";
+import Header from "@/components/global/header";
 import { Toaster } from "@/components/ui/sonner";
 import { ThemeProvider } from "@/providers/theme-provider";
 import { esMX } from "@clerk/localizations";
